src/test/java/com/sap/gs/techcounsil/controller/DemoControllerTests.java
package com.sap.gs.techcounsil.controller;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

@RunWith(SpringRunner.class)
@WebMvcTest(DemoController.class)
public class DemoControllerTests {

	@Autowired
	private MockMvc mockMvc;

	@Test
	public void callHelloWorldV1() throws Exception {
		mockMvc.perform(get("/v1/helloWorld")).andExpect(status().isOk())
				.andExpect(content().contentType(MediaType.APPLICATION_JSON_UTF8_VALUE))
				.andExpect(content().string("{\"hello\": \"world\"}"));
	}

}
